requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
Babel==2.8.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.3.9
Delorean==1.0.0
docutils==0.15.2
Flask==1.1.1
flask-restx==0.1.1
Flask-SQLAlchemy==2.4.1
future==0.17.1
gunicorn==20.0.4
humanize==2.0.0
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.0.0b9
pathspec==0.7.0
pyasn1==0.4.7
pyrsistent==0.15.7
python-dateutil==2.8.0
python-dotenv==0.12.0
pytz==2019.2
PyYAML==5.1.2
regex==2020.2.20
requests==2.23.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.15
sqlparse==0.3.0
supervisor==4.1.0
toml==0.10.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.3
uuid==1.30
virtualenv==16.7.7
Werkzeug==0.16.1
zipp==3.1.0
